requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-replace==2020.12.3
django-rosetta==0.9.9
idna==3.4
Pillow==10.0.0
polib==1.2.0
psycopg2==2.9.7
python-slugify==8.0.1
redis==5.0.0
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.4
